requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.2.0
langgraph>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
crewai>=0.80.0
